fix(ui): fix storybook types
